fix(memory): release unused size when purging app
